The data model has been changed. The report table is divided into a report table and a report relationship table with a category. Implemented moving, deleting categories according to the new data model. Implemented the ability to change the parent category of reports.

+++ ReportingAPI/Controllers/ReportsController.cs
@@ -27,19 +27,33 @@ namespace ReportingApi.Controllers
             _mapper = mapper;
         }
 
+
         [AllowAnonymous]
         // GET: api/Reports
         [HttpGet]
         public async Task<List<Report>> GetReports()
         {
-            return await _context.Reports.ToListAsync();
-        }
+            List<Report> reports = await _context.Reports.Include(x => x.Categories).AsNoTracking().ToListAsync();
+            List<Report> outRep = new();
 
+            foreach(var report in reports)
+                foreach (var category in report.Categories)
+                {
+                    Report temp = report.Clone();
+                    temp.ParentId = category.Id;
+                    outRep.Add(temp);
+                }
+
+                    return outRep;
+           //     ToListAsync();
+        }
+        
         // PUT: api/PutReport
         [HttpPut]
         public async Task<ActionResult> PutReport(UpdateReport report)
         {
-            Report Reports = _mapper.Map<Report>(report);
+            //Report Reports = _mapper.Map<Report>(report);
+            Report Reports = await _context.Reports.FirstOrDefaultAsync(x => x.Id == report.Id);
             var results = new List<ValidationResult>();
             var context = new System.ComponentModel.DataAnnotations.ValidationContext(report);
             if (!Validator.TryValidateObject(report, context, results, true))
@@ -49,6 +63,25 @@ namespace ReportingApi.Controllers
                     Console.WriteLine(error.ErrorMessage);
                 }
             }
+            var DuplicateByText = _context.Reports.Where(x => x.Text == report.Text && x.Id != report.Id);
+            var hasDuplicateByText = DuplicateByText.Count(x => x.Text == report.Text && x.Id != report.Id) > 0;
+            var DuplicateByURL = _context.Reports.Where(x => x.URL == report.URL && x.Id != report.Id);
+            var hasDuplicateByURL = DuplicateByURL.Count(x => x.URL == report.URL && x.Id != report.Id) > 0;
+            string Alias = "";
+
+            if (hasDuplicateByText)
+            {
+                Alias = DuplicateByText.First().Alias;
+                return BadRequest($"Отчет с таким именем уже существует.\nПсевдоним отчета: {Alias}");
+            }
+            else if (hasDuplicateByURL)
+            {
+                Alias = DuplicateByURL.First().Alias;
+                return BadRequest($"Отчет с такой ссылкой уже существует.\nПсевдоним отчета: {Alias}");
+            }
+ 
+            Reports.Text = report.Text;
+            Reports.URL = report.URL;
             _context.Entry(Reports).State = EntityState.Modified;
 
             try
@@ -63,12 +96,32 @@ namespace ReportingApi.Controllers
         }
 
         // PUT: api/PutParentId
-        [HttpPut("{id}")]
-        public async Task<ActionResult> PutParentId(int id, [FromBody] int? parentId)
+        //[AllowAnonymous]
+        [HttpPut("UpdateCategoryReports")]
+        ///api/Reports/UpdateCategoryReports
+        public async Task<ActionResult> PutParentId(UpdateCategoryReports CategoryReports = null)
         {
-            var report = await _context.Reports.FindAsync(id);
-            report.ParentId = parentId;
-            _context.Reports.Update(report);
+            Report report = _context.Reports.Include(x => x.Categories).FirstOrDefault(x => x.Id == CategoryReports.id && x.Categories.Any(y => y.Id == CategoryReports.fromCat));
+            //Console.WriteLine("Test: " + report.Categories.Count());
+            Category fromCategory = report.Categories.FirstOrDefault(y => y.Id == CategoryReports.fromCat);
+            if (report == null || fromCategory == null)
+            {
+                return BadRequest("Отчета с данным Id в указанной категории не существует");
+            }
+            Console.WriteLine(CategoryReports.toCat);
+            var HasDublicateInCat = report.Categories.Any(x => x.Id == CategoryReports.toCat);
+
+            if (HasDublicateInCat)
+            {
+                return BadRequest("В данной категории уже существует такой отчет. Размещение одинаковых отчетов в одной категории запрещено.");
+            }
+
+            //var t = _context.Reports.Include(x => x.Categories).AsNoTracking().FirstOrDefault(x => CategoryReports.id == x.Id && x.Categories.Any(y => y.Id == CategoryReports.toCat));
+            //Console.WriteLine();
+            report.Categories.Remove(fromCategory);
+
+            Category toCategory = _context.Categories.Include(x => x.Reports).FirstOrDefault(y => y.Id == CategoryReports.toCat);
+            report.Categories.Add(toCategory);
 
             try
             {
@@ -88,6 +141,14 @@ namespace ReportingApi.Controllers
             Report report = _mapper.Map<Report>(newReport);
             try
             {
+                Category category = _context.Categories.FirstOrDefault(x => x.Id == newReport.ParentId);
+                if(category == null)
+                {
+                    return BadRequest("Категория не существует.");
+                }
+
+                report.Categories.Add(category);
+
                 _context.Reports.Add(report);
                 await _context.SaveChangesAsync();
             }
@@ -106,10 +167,23 @@ namespace ReportingApi.Controllers
         {
             var report = await _context.Reports.FindAsync(id);
 
+
+            /*Report report = _context.Reports.Include(x => x.Categories).FirstOrDefault(x => x.Id == CategoryReports.id && x.Categories.Any(y => y.Id == CategoryReports.fromCat));
+            Console.WriteLine("Test: " + report.Categories.Count());
+            Category fromCategory = report.Categories.FirstOrDefault(y => y.Id == CategoryReports.fromCat);
+            if (report == null || fromCategory == null)
+            {
+                throw new Exception("Отчета с данным Id в указанной категории не существует");
+            }
+            report.Categories.Remove(fromCategory);
+
+            Category toCategory = _context.Categories.Include(x => x.Reports).FirstOrDefault(y => y.Id == CategoryReports.toCat);
+            report.Categories.Add(toCategory);*/
+
             try
             {
                 _context.Reports.Remove(report);
-                await _context.SaveChangesAsync();
+              //  await _context.SaveChangesAsync();
                 return Ok();
             }
             catch (Exception e)
